Relin key gen (#3)
* WIP - Relin Key Gen first pass no tests

* WIP - Relin Key Gen tests are failing

* Formally disallow ct.len > 2

M-BFV does not support this and it is non-trivial to adapt (can't just square keys because the sum of squares will not equal the square of the sum)

* Add test for homomorphic addition

* Fix invalid representation

* Finish RLK gen

Oddly, all of this fails unless the PK CRP coefficients are "smaller". I
guess encryption brings in a bit too much noise otherwise. Not sure.

I also tried instantiating secret keys with ternary coefficients, but
still the tests would fail unless the PK CRP coefficients were generated
smaller.