Multipary BFV (#218)
* Encrypt & decrypt (#1)

* Implement public key gen

* Avoid visibility changes to fhe crate

And just stick threshold into a module for now

* Implement threshold encrypt/decrypt

* Implement public key switch (#2)

* Relin key gen (#3)

* WIP - Relin Key Gen first pass no tests

* WIP - Relin Key Gen tests are failing

* Formally disallow ct.len > 2

M-BFV does not support this and it is non-trivial to adapt (can't just square keys because the sum of squares will not equal the square of the sum)

* Add test for homomorphic addition

* Fix invalid representation

* Finish RLK gen

Oddly, all of this fails unless the PK CRP coefficients are "smaller". I
guess encryption brings in a bit too much noise otherwise. Not sure.

I also tried instantiating secret keys with ternary coefficients, but
still the tests would fail unless the PK CRP coefficients were generated
smaller.

* Slight improvements to the API (#4)

* Voting example (#5)

* Add voting example

* Fix module structure and API docs

* Wrap the CRP in its own type

The main fhe library does not expose the underlying `fhe_math::Poly` type in any of its API, so we shouldn't either. Also this is a bit cleaner for library consumers.

* Fix multiparty relin key (#6)

* Clean up comments / unused code

* Fix CI failures

* Address PR feedback

Allow some variable time computations and include link to the paper